+++ frappe/search/__init__.py
@@ -1,0 +1,13 @@
+# Copyright (c) 2020, Frappe Technologies Pvt. Ltd. and Contributors
+# MIT License. See license.txt
+
+import frappe
+from frappe.utils import cint
+from frappe.search.website_search import WebsiteSearch
+from frappe.search.website_search import FullTextSearch
+
+@frappe.whitelist(allow_guest=True)
+def web_search(query, scope=None, limit=20):
+	limit = cint(limit)
+	ws = WebsiteSearch(index_name="web_routes")
+	return ws.search(query, scope, limit)
